+++ app/src/main/java/com/freshdigitable/udonroad2/di/MainActivityModule.kt
@@ -18,19 +18,16 @@ package com.freshdigitable.udonroad2.di
 
 import androidx.appcompat.app.AppCompatActivity
 import androidx.lifecycle.ViewModel
-import com.freshdigitable.udonroad2.data.impl.TweetRepository
 import com.freshdigitable.udonroad2.input.di.TweetInputFragmentModule
 import com.freshdigitable.udonroad2.main.DrawerViewModelSource
 import com.freshdigitable.udonroad2.main.MainActivity
 import com.freshdigitable.udonroad2.main.MainActivityNavigationDelegate
 import com.freshdigitable.udonroad2.main.MainViewModel
 import com.freshdigitable.udonroad2.main.MainViewModelSource
-import com.freshdigitable.udonroad2.model.ListOwnerGenerator
 import com.freshdigitable.udonroad2.model.app.di.ViewModelKey
 import com.freshdigitable.udonroad2.model.app.navigation.ActivityEffectDelegate
 import com.freshdigitable.udonroad2.model.app.navigation.EventDispatcher
 import com.freshdigitable.udonroad2.settings.di.AppSettingFragmentModule
-import com.freshdigitable.udonroad2.shortcut.ShortcutActions
 import com.freshdigitable.udonroad2.shortcut.ShortcutViewStates
 import dagger.Binds
 import dagger.Module
@@ -65,16 +62,5 @@ internal interface MainActivityModule {
             shortcutViewStates: ShortcutViewStates,
         ): ViewModel =
             MainViewModel(navigator, viewState, drawerViewModelSource, shortcutViewStates)
-
-        @Provides
-        fun provideShortcutViewStates(
-            dispatcher: EventDispatcher,
-            tweetRepository: TweetRepository,
-            listOwnerGenerator: ListOwnerGenerator,
-        ): ShortcutViewStates {
-            return ShortcutViewStates.create(
-                ShortcutActions.create(dispatcher), tweetRepository, listOwnerGenerator
-            )
-        }
     }
 }
