+++ feature/shortcut/src/main/java/com/freshdigitable/udonroad2/shortcut/ShortcutActions.kt
@@ -16,40 +16,71 @@
 
 package com.freshdigitable.udonroad2.shortcut
 
+import android.view.MenuItem
+import com.freshdigitable.udonroad2.model.SelectedItemId
+import com.freshdigitable.udonroad2.model.TweetId
 import com.freshdigitable.udonroad2.model.app.navigation.EventDispatcher
+import com.freshdigitable.udonroad2.model.app.navigation.postEvents
 import com.freshdigitable.udonroad2.model.app.navigation.toActionFlow
-import kotlinx.coroutines.flow.Flow
+import javax.inject.Inject
 
-interface ShortcutActions : ShortcutEventListener {
-    val showTweetDetail: Flow<SelectedItemShortcut.TweetDetail>
-    val favTweet: Flow<SelectedItemShortcut.Like>
-    val retweet: Flow<SelectedItemShortcut.Retweet>
-    val showConversation: Flow<SelectedItemShortcut.Conversation>
-    val unlikeTweet: Flow<SelectedItemShortcut.Unlike>
-    val unretweetTweet: Flow<SelectedItemShortcut.Unretweet>
-    val deleteTweet: Flow<SelectedItemShortcut.DeleteTweet>
+class ShortcutActions @Inject constructor(
+    private val dispatcher: EventDispatcher,
+) : ShortcutEventListener {
+    internal val showTweetDetail = dispatcher.toActionFlow<SelectedItemShortcut.TweetDetail>()
+    internal val favTweet = dispatcher.toActionFlow<SelectedItemShortcut.Like>()
+    internal val retweet = dispatcher.toActionFlow<SelectedItemShortcut.Retweet>()
+    internal val showConversation = dispatcher.toActionFlow<SelectedItemShortcut.Conversation>()
+    internal val unlikeTweet = dispatcher.toActionFlow<SelectedItemShortcut.Unlike>()
+    internal val unretweetTweet = dispatcher.toActionFlow<SelectedItemShortcut.Unretweet>()
+    internal val deleteTweet = dispatcher.toActionFlow<SelectedItemShortcut.DeleteTweet>()
 
-    companion object {
-        fun create(
-            dispatcher: EventDispatcher,
-            listener: ShortcutEventListener = ShortcutEventListener.create(dispatcher),
-        ): ShortcutActions = TweetShortcutActions(dispatcher, listener)
+    override fun onShortcutMenuSelected(item: MenuItem, id: TweetId) {
+        dispatcher.postSelectedItemShortcutEvent(item, id)
     }
 }
 
-private class TweetShortcutActions(
-    dispatcher: EventDispatcher,
-    listener: ShortcutEventListener,
-) : ShortcutActions, ShortcutEventListener by listener {
-    override val showTweetDetail: Flow<SelectedItemShortcut.TweetDetail> =
-        dispatcher.toActionFlow()
-    override val favTweet: Flow<SelectedItemShortcut.Like> =
-        dispatcher.toActionFlow()
-    override val retweet: Flow<SelectedItemShortcut.Retweet> =
-        dispatcher.toActionFlow()
-    override val showConversation: Flow<SelectedItemShortcut.Conversation> =
-        dispatcher.toActionFlow()
-    override val unlikeTweet = dispatcher.toActionFlow<SelectedItemShortcut.Unlike>()
-    override val unretweetTweet = dispatcher.toActionFlow<SelectedItemShortcut.Unretweet>()
-    override val deleteTweet = dispatcher.toActionFlow<SelectedItemShortcut.DeleteTweet>()
+fun EventDispatcher.postSelectedItemShortcutEvent(
+    menuItem: MenuItem,
+    selectedItemId: SelectedItemId,
+) {
+    val tweetId = selectedItemId.quoteId ?: selectedItemId.originalId
+    postSelectedItemShortcutEvent(menuItem, tweetId)
+}
+
+fun EventDispatcher.postSelectedItemShortcutEvent(
+    menuItem: MenuItem,
+    tweetId: TweetId,
+) {
+    when (menuItem.itemId) {
+        R.id.iffabMenu_main_detail -> postEvent(SelectedItemShortcut.TweetDetail(tweetId))
+        R.id.iffabMenu_main_fav -> postEvent(SelectedItemShortcut.Like(tweetId))
+        R.id.iffabMenu_main_rt -> postEvent(SelectedItemShortcut.Retweet(tweetId))
+        R.id.iffabMenu_main_favRt -> postEvents(
+            SelectedItemShortcut.Like(tweetId),
+            SelectedItemShortcut.Retweet(tweetId)
+        )
+        R.id.iffabMenu_main_reply -> postEvent(SelectedItemShortcut.Reply(tweetId))
+        R.id.iffabMenu_main_quote -> postEvent(SelectedItemShortcut.Quote(tweetId))
+        R.id.iffabMenu_main_conv -> postEvent(SelectedItemShortcut.Conversation(tweetId))
+        R.id.detail_main_rt -> {
+            if (menuItem.isChecked) {
+                postEvent(SelectedItemShortcut.Unretweet(tweetId))
+            } else {
+                postEvent(SelectedItemShortcut.Retweet(tweetId))
+            }
+        }
+        R.id.detail_main_fav -> {
+            if (menuItem.isChecked) {
+                postEvent(SelectedItemShortcut.Unlike(tweetId))
+            } else {
+                postEvent(SelectedItemShortcut.Like(tweetId))
+            }
+        }
+        R.id.detail_main_reply -> postEvent(SelectedItemShortcut.Reply(tweetId))
+        R.id.detail_main_quote -> postEvent(SelectedItemShortcut.Quote(tweetId))
+        R.id.detail_more_delete -> postEvent(SelectedItemShortcut.DeleteTweet(tweetId))
+        R.id.detail_main_conv -> postEvent(SelectedItemShortcut.Conversation(tweetId))
+        else -> throw IllegalStateException("selected unregistered menu item: $menuItem")
+    }
 }
